refactor(utils): rename file with util
