+++ backend/src/middlewares/error.middleware.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { HttpException } from "../exceptions/http.exception";
-import { logger } from "../utils/logger";
+import { logger } from "../utils/logger.util";
 
 export const errorMiddleware = (
   error: HttpException,
